inline all actions*.py into alpamayo_r1.py, remove dataset.py
